refactor: reorganize
